# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Staffing Platform API Doc
  description: |
    On-Demand Healthcare Staffing & Visit Management Portal — REST API.

    ## Authentication
    All protected endpoints require a **Bearer JWT** in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```
    Obtain tokens via `POST /auth/login`. Access tokens expire in **15 minutes**;
    use `POST /auth/refresh` with a valid refresh token to rotate.

    ## Roles
    | Role | Description |
    |------|-------------|
    | `SUPER_ADMIN` | Full platform control |
    | `RECRUITER` | Read-only admin access |
    | `FACILITY_ADMIN` | Manages one facility tenant |
    | `TEAM_MEMBER` | Scoped facility operator |
    | `NURSE` | Field professional (mobile) |

    ## File Uploads
    All uploads are stored in **DigitalOcean Spaces** (private).
    Use `POST /storage/signed-upload-url` for direct client-to-Spaces uploads,
    or `POST /storage/upload/{folder}` for server-proxied multipart uploads.
    Download links are short-lived signed URLs — never public.

    ## Pagination
    List endpoints accept `page` and `limit` query params and return:
    ```json
    { "pagination": { "page": 1, "limit": 20, "total": 100, "totalPages": 5 } }
    ```

    ## EVV Geofencing
    Check-in/out enforces a **200-metre geofence** around the case address (Haversine).
    Breaches create a `FLAGGED` visit requiring admin override approval.
  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable Parameters ──────────────────
...
  responses:
...
    Forbidden:
      description: Authenticated but insufficient role/scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    StorageFolder:
      type: string
      enum: [credentials, facility-logos, facility-documents, avatars,
             chat-attachments, visit-signatures, admin-documents]

    # ── Generic Wrappers ───────────────────
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data:    { type: object }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: 'Resource not found' }
...
    UploadResult:
      type: object
      properties:
        objectKey:    { type: string, example: 'credentials/abc123.pdf' }
        originalName: { type: string }
        size:         { type: integer, description: 'Bytes' }
        mimeType:     { type: string }
        downloadUrl:  { type: string, description: 'Short-lived signed URL' }
        expiresIn:    { type: integer, description: 'Seconds until URL expires' }
security:
  - BearerAuth: []

# docs/paths-part1.yaml  (auth · users · credentials · facilities)
paths:
...
  /storage/signed-upload-url:
    post:
      tags: [Storage]
      summary: Request a pre-signed PUT URL for direct-to-Spaces upload
      description: |
        Preferred for large files. The client PUTs the file directly to DigitalOcean Spaces
        using the returned URL (valid for **5 minutes**).  
        After a successful upload, call `POST /storage/confirm` to validate and get a download URL.

        **Folder access is role-gated:**

        | Folder | Allowed Roles |
        |--------|---------------|
        | `credentials` | NURSE |
        | `facility-logos` | SUPER_ADMIN, FACILITY_ADMIN |
        | `facility-documents` | SUPER_ADMIN, FACILITY_ADMIN, TEAM_MEMBER |
        | `avatars` | All roles |
        | `chat-attachments` | All roles |
        | `visit-signatures` | NURSE |
        | `admin-documents` | SUPER_ADMIN |
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [folder, filename, contentType]
              properties:
                folder:      { $ref: '#/components/schemas/StorageFolder' }
                filename:    { type: string, example: 'nursing-license.pdf' }
                contentType: { type: string, example: 'application/pdf' }
      responses:
        '201':
          description: Signed URL generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          uploadUrl:   { type: string, description: 'Pre-signed PUT URL (5 min)' }
                          objectKey:   { type: string, example: 'credentials/abc-uuid.pdf' }
                          expiresIn:   { type: integer, example: 300 }
                          maxBytes:    { type: integer }
                          confirmUrl:  { type: string }
        '403': { $ref: '#/components/responses/Forbidden' }
        '415':
          description: File type not allowed for this folder
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /storage/confirm:
    post:
      tags: [Storage]
      summary: Confirm a completed direct upload and receive a download URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objectKey]
              properties:
                objectKey: { type: string, example: 'credentials/abc-uuid.pdf' }
      responses:
        '200':
          description: Upload confirmed, download URL returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          objectKey:   { type: string }
                          downloadUrl: { type: string, description: 'Signed URL valid for 1 hour' }
                          expiresIn:   { type: integer, example: 3600 }
        '403': { $ref: '#/components/responses/Forbidden' }
  /storage/upload/{folder}:
    post:
      tags: [Storage]
      summary: Proxied multipart upload through the API server
      description: |
        Suitable for smaller files. The file is streamed through the API to Spaces.  
        For large files use `POST /storage/signed-upload-url` instead.
      parameters:
        - name: folder
          in: path
          required: true
          schema: { $ref: '#/components/schemas/StorageFolder' }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/UploadResult' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413':
          description: File exceeds folder size limit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '415':
          description: File type not allowed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /storage/download:
    get:
      tags: [Storage]
      summary: Generate a fresh signed download URL for a private object
      parameters:
        - name: key
          in: query
          required: true
          description: Spaces object key (e.g. `credentials/abc.pdf`)
          schema: { type: string }
        - name: expiresIn
          in: query
          description: URL validity in seconds (max 86400 / 24 h)
          schema: { type: integer, default: 3600 }
      responses:
        '200':
          description: Signed download URL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          objectKey:   { type: string }
                          downloadUrl: { type: string }
                          expiresIn:   { type: integer }
        '403': { $ref: '#/components/responses/Forbidden' }
  /storage:
    delete:
      tags: [Storage]
      summary: Hard-delete an object from Spaces
      description: '**Roles:** SUPER_ADMIN'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key: { type: string, example: 'credentials/abc.pdf' }
      responses:
        '200':
          description: Object deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          objectKey: { type: string }
        '403': { $ref: '#/components/responses/Forbidden' }